Initial Commit of Project
This is my take on the Space Shooter tutorial from the Unity website
written in C#. There are a lot of features that I would like to add in
order to further develop my Unity skills. Some of these features include
Hazards, Power-ups, multiple levels and player customization. The
purpose of this commit is to showcase the completed tutorial.